requirements.txt
torch
transformers
openai-whisper
opencv-python
scenedetect
Pillow
accelerate
ffmpeg-python
fastapi
uvicorn
python-multipart
jinja2
moviepy
